api/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pydantic==2.8.2
pydantic-settings==2.3.4
httpx==0.27.0
sqlalchemy==2.0.32
python-dotenv==1.0.1
# Requires Responses API support. Upgrade with: pip install -U openai
openai>=1.40.0
reportlab==4.2.2
pypdf==5.2.0
